+++ src/snowflake/snowpark/_internal/analyzer/analyzer.py
@@ -3,8 +3,8 @@
 # Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
 #
 
-from collections import Counter
-from typing import Dict, Union
+from collections import Counter, defaultdict
+from typing import DefaultDict, Dict, Union
 
 import snowflake.snowpark
 from snowflake.snowpark._internal.analyzer.analyzer_utils import (
@@ -154,23 +154,42 @@ class Analyzer:
         self.alias_maps_to_use = None
 
     def analyze(
-        self, expr: Union[Expression, NamedExpression], parse_local_name=False
+        self,
+        expr: Union[Expression, NamedExpression],
+        df_aliased_col_name_to_real_col_name: DefaultDict[str, Dict[str, str]],
+        parse_local_name=False,
     ) -> str:
         if isinstance(expr, GroupingSetsExpression):
             return grouping_set_expression(
-                [[self.analyze(a, parse_local_name) for a in arg] for arg in expr.args]
+                [
+                    [
+                        self.analyze(
+                            a, df_aliased_col_name_to_real_col_name, parse_local_name
+                        )
+                        for a in arg
+                    ]
+                    for arg in expr.args
+                ]
             )
 
         if isinstance(expr, Like):
             return like_expression(
-                self.analyze(expr.expr, parse_local_name),
-                self.analyze(expr.pattern, parse_local_name),
+                self.analyze(
+                    expr.expr, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                self.analyze(
+                    expr.pattern, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
             )
 
         if isinstance(expr, RegExp):
             return regexp_expression(
-                self.analyze(expr.expr, parse_local_name),
-                self.analyze(expr.pattern, parse_local_name),
+                self.analyze(
+                    expr.expr, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                self.analyze(
+                    expr.pattern, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
             )
 
         if isinstance(expr, Collate):
@@ -178,25 +197,45 @@ class Analyzer:
                 expr.collation_spec.upper() if parse_local_name else expr.collation_spec
             )
             return collate_expression(
-                self.analyze(expr.expr, parse_local_name), collation_spec
+                self.analyze(
+                    expr.expr, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                collation_spec,
             )
 
         if isinstance(expr, (SubfieldString, SubfieldInt)):
             field = expr.field
             if parse_local_name and isinstance(field, str):
                 field = field.upper()
-            return subfield_expression(self.analyze(expr.expr, parse_local_name), field)
+            return subfield_expression(
+                self.analyze(
+                    expr.expr, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                field,
+            )
 
         if isinstance(expr, CaseWhen):
             return case_when_expression(
                 [
                     (
-                        self.analyze(condition, parse_local_name),
-                        self.analyze(value, parse_local_name),
+                        self.analyze(
+                            condition,
+                            df_aliased_col_name_to_real_col_name,
+                            parse_local_name,
+                        ),
+                        self.analyze(
+                            value,
+                            df_aliased_col_name_to_real_col_name,
+                            parse_local_name,
+                        ),
                     )
                     for condition, value in expr.branches
                 ],
-                self.analyze(expr.else_value, parse_local_name)
+                self.analyze(
+                    expr.else_value,
+                    df_aliased_col_name_to_real_col_name,
+                    parse_local_name,
+                )
                 if expr.else_value
                 else "NULL",
             )
@@ -204,39 +243,79 @@ class Analyzer:
         if isinstance(expr, MultipleExpression):
             return block_expression(
                 [
-                    self.analyze(expression, parse_local_name)
+                    self.analyze(
+                        expression,
+                        df_aliased_col_name_to_real_col_name,
+                        parse_local_name,
+                    )
                     for expression in expr.expressions
                 ]
             )
 
         if isinstance(expr, InExpression):
             return in_expression(
-                self.analyze(expr.columns, parse_local_name),
+                self.analyze(
+                    expr.columns, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
                 [
-                    self.analyze(expression, parse_local_name)
+                    self.analyze(
+                        expression,
+                        df_aliased_col_name_to_real_col_name,
+                        parse_local_name,
+                    )
                     for expression in expr.values
                 ],
             )
 
         if isinstance(expr, GroupingSet):
-            return self.grouping_extractor(expr)
+            return self.grouping_extractor(expr, df_aliased_col_name_to_real_col_name)
 
         if isinstance(expr, WindowExpression):
             return window_expression(
-                self.analyze(expr.window_function, parse_local_name),
-                self.analyze(expr.window_spec, parse_local_name),
+                self.analyze(
+                    expr.window_function,
+                    df_aliased_col_name_to_real_col_name,
+                    parse_local_name,
+                ),
+                self.analyze(
+                    expr.window_spec,
+                    df_aliased_col_name_to_real_col_name,
+                    parse_local_name,
+                ),
             )
         if isinstance(expr, WindowSpecDefinition):
             return window_spec_expression(
-                [self.analyze(x, parse_local_name) for x in expr.partition_spec],
-                [self.analyze(x, parse_local_name) for x in expr.order_spec],
-                self.analyze(expr.frame_spec, parse_local_name),
+                [
+                    self.analyze(
+                        x, df_aliased_col_name_to_real_col_name, parse_local_name
+                    )
+                    for x in expr.partition_spec
+                ],
+                [
+                    self.analyze(
+                        x, df_aliased_col_name_to_real_col_name, parse_local_name
+                    )
+                    for x in expr.order_spec
+                ],
+                self.analyze(
+                    expr.frame_spec,
+                    df_aliased_col_name_to_real_col_name,
+                    parse_local_name,
+                ),
             )
         if isinstance(expr, SpecifiedWindowFrame):
             return specified_window_frame_expression(
                 expr.frame_type.sql,
-                self.window_frame_boundary(self.to_sql_avoid_offset(expr.lower)),
-                self.window_frame_boundary(self.to_sql_avoid_offset(expr.upper)),
+                self.window_frame_boundary(
+                    self.to_sql_avoid_offset(
+                        expr.lower, df_aliased_col_name_to_real_col_name
+                    )
+                ),
+                self.window_frame_boundary(
+                    self.to_sql_avoid_offset(
+                        expr.upper, df_aliased_col_name_to_real_col_name
+                    )
+                ),
             )
         if isinstance(expr, UnspecifiedFrame):
             return ""
@@ -254,6 +333,15 @@ class Analyzer:
             return quote_name(name)
 
         if isinstance(expr, UnresolvedAttribute):
+            if expr.df_alias:
+                if expr.df_alias in df_aliased_col_name_to_real_col_name:
+                    return df_aliased_col_name_to_real_col_name[expr.df_alias].get(
+                        expr.name, expr.name
+                    )
+                else:
+                    raise SnowparkClientExceptionMessages.DF_ALIAS_NOT_RECOGNIZED(
+                        expr.df_alias
+                    )
             return expr.name
 
         if isinstance(expr, FunctionExpression):
@@ -264,15 +352,32 @@ class Analyzer:
             func_name = expr.name.upper() if parse_local_name else expr.name
             return function_expression(
                 func_name,
-                [self.to_sql_avoid_offset(c) for c in expr.children],
+                [
+                    self.to_sql_avoid_offset(c, df_aliased_col_name_to_real_col_name)
+                    for c in expr.children
+                ],
                 expr.is_distinct,
             )
 
         if isinstance(expr, Star):
+            if expr.df_alias:
+                # This is only hit by col(<df_alias>)
+                if expr.df_alias not in df_aliased_col_name_to_real_col_name:
+                    raise SnowparkClientExceptionMessages.DF_ALIAS_NOT_RECOGNIZED(
+                        expr.df_alias
+                    )
+                columns = df_aliased_col_name_to_real_col_name[expr.df_alias]
+                return ",".join(columns.values())
             if not expr.expressions:
                 return "*"
             else:
-                return ",".join(list(map(self.analyze, expr.expressions)))
+                # This case is hit by df.col("*")
+                return ",".join(
+                    [
+                        self.analyze(e, df_aliased_col_name_to_real_col_name)
+                        for e in expr.expressions
+                    ]
+                )
 
         if isinstance(expr, SnowflakeUDF):
             if expr.api_call_source is not None:
@@ -282,7 +387,12 @@ class Analyzer:
             func_name = expr.udf_name.upper() if parse_local_name else expr.udf_name
             return function_expression(
                 func_name,
-                [self.analyze(x, parse_local_name) for x in expr.children],
+                [
+                    self.analyze(
+                        x, df_aliased_col_name_to_real_col_name, parse_local_name
+                    )
+                    for x in expr.children
+                ],
                 False,
             )
 
@@ -291,25 +401,41 @@ class Analyzer:
                 self.session._conn._telemetry_client.send_function_usage_telemetry(
                     expr.api_call_source, TelemetryField.FUNC_CAT_USAGE.value
                 )
-            return self.table_function_expression_extractor(expr)
+            return self.table_function_expression_extractor(
+                expr, df_aliased_col_name_to_real_col_name
+            )
 
         if isinstance(expr, TableFunctionPartitionSpecDefinition):
             return table_function_partition_spec(
                 expr.over,
-                [self.analyze(x, parse_local_name) for x in expr.partition_spec]
+                [
+                    self.analyze(
+                        x, df_aliased_col_name_to_real_col_name, parse_local_name
+                    )
+                    for x in expr.partition_spec
+                ]
                 if expr.partition_spec
                 else [],
-                [self.analyze(x, parse_local_name) for x in expr.order_spec]
+                [
+                    self.analyze(
+                        x, df_aliased_col_name_to_real_col_name, parse_local_name
+                    )
+                    for x in expr.order_spec
+                ]
                 if expr.order_spec
                 else [],
             )
 
         if isinstance(expr, UnaryExpression):
-            return self.unary_expression_extractor(expr, parse_local_name)
+            return self.unary_expression_extractor(
+                expr, df_aliased_col_name_to_real_col_name, parse_local_name
+            )
 
         if isinstance(expr, SortOrder):
             return order_expression(
-                self.analyze(expr.child, parse_local_name),
+                self.analyze(
+                    expr.child, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
                 expr.direction.sql,
                 expr.null_ordering.sql,
             )
@@ -320,34 +446,60 @@ class Analyzer:
 
         if isinstance(expr, WithinGroup):
             return within_group_expression(
-                self.analyze(expr.expr, parse_local_name),
-                [self.analyze(e) for e in expr.order_by_cols],
+                self.analyze(
+                    expr.expr, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                [
+                    self.analyze(e, df_aliased_col_name_to_real_col_name)
+                    for e in expr.order_by_cols
+                ],
             )
 
         if isinstance(expr, BinaryExpression):
-            return self.binary_operator_extractor(expr, parse_local_name)
+            return self.binary_operator_extractor(
+                expr, df_aliased_col_name_to_real_col_name, parse_local_name
+            )
 
         if isinstance(expr, InsertMergeExpression):
             return insert_merge_statement(
-                self.analyze(expr.condition) if expr.condition else None,
-                [self.analyze(k) for k in expr.keys],
-                [self.analyze(v) for v in expr.values],
+                self.analyze(expr.condition, df_aliased_col_name_to_real_col_name)
+                if expr.condition
+                else None,
+                [
+                    self.analyze(k, df_aliased_col_name_to_real_col_name)
+                    for k in expr.keys
+                ],
+                [
+                    self.analyze(v, df_aliased_col_name_to_real_col_name)
+                    for v in expr.values
+                ],
             )
 
         if isinstance(expr, UpdateMergeExpression):
             return update_merge_statement(
-                self.analyze(expr.condition) if expr.condition else None,
-                {self.analyze(k): self.analyze(v) for k, v in expr.assignments.items()},
+                self.analyze(expr.condition, df_aliased_col_name_to_real_col_name)
+                if expr.condition
+                else None,
+                {
+                    self.analyze(k, df_aliased_col_name_to_real_col_name): self.analyze(
+                        v, df_aliased_col_name_to_real_col_name
+                    )
+                    for k, v in expr.assignments.items()
+                },
             )
 
         if isinstance(expr, DeleteMergeExpression):
             return delete_merge_statement(
-                self.analyze(expr.condition) if expr.condition else None
+                self.analyze(expr.condition, df_aliased_col_name_to_real_col_name)
+                if expr.condition
+                else None
             )
 
         if isinstance(expr, ListAgg):
             return list_agg(
-                self.analyze(expr.col, parse_local_name),
+                self.analyze(
+                    expr.col, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
                 str_to_sql(expr.delimiter),
                 expr.is_distinct,
             )
@@ -355,9 +507,15 @@ class Analyzer:
         if isinstance(expr, RankRelatedFunctionExpression):
             return rank_related_function_expression(
                 expr.sql,
-                self.analyze(expr.expr, parse_local_name),
+                self.analyze(
+                    expr.expr, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
                 expr.offset,
-                self.analyze(expr.default, parse_local_name) if expr.default else None,
+                self.analyze(
+                    expr.default, df_aliased_col_name_to_real_col_name, parse_local_name
+                )
+                if expr.default
+                else None,
                 expr.ignore_nulls,
             )
 
@@ -366,11 +524,16 @@ class Analyzer:
         )  # pragma: no cover
 
     def table_function_expression_extractor(
-        self, expr: TableFunctionExpression, parse_local_name=False
+        self,
+        expr: TableFunctionExpression,
+        df_aliased_col_name_to_real_col_name: Dict[str, str],
+        parse_local_name=False,
     ) -> str:
         if isinstance(expr, FlattenFunction):
             return flatten_expression(
-                self.analyze(expr.input, parse_local_name),
+                self.analyze(
+                    expr.input, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
                 expr.path,
                 expr.outer,
                 expr.recursive,
@@ -379,14 +542,21 @@ class Analyzer:
         elif isinstance(expr, PosArgumentsTableFunction):
             sql = function_expression(
                 expr.func_name,
-                [self.analyze(x, parse_local_name) for x in expr.args],
+                [
+                    self.analyze(
+                        x, df_aliased_col_name_to_real_col_name, parse_local_name
+                    )
+                    for x in expr.args
+                ],
                 False,
             )
         elif isinstance(expr, (NamedArgumentsTableFunction, GeneratorTableFunction)):
             sql = named_arguments_function(
                 expr.func_name,
                 {
-                    key: self.analyze(value, parse_local_name)
+                    key: self.analyze(
+                        value, df_aliased_col_name_to_real_col_name, parse_local_name
+                    )
                     for key, value in expr.args.items()
                 },
             )
@@ -396,12 +566,17 @@ class Analyzer:
                 "NamedArgumentsTableFunction, GeneratorTableFunction, or FlattenFunction."
             )
         partition_spec_sql = (
-            self.analyze(expr.partition_spec) if expr.partition_spec else ""
+            self.analyze(expr.partition_spec, df_aliased_col_name_to_real_col_name)
+            if expr.partition_spec
+            else ""
         )
         return f"{sql} {partition_spec_sql}"
 
     def unary_expression_extractor(
-        self, expr: UnaryExpression, parse_local_name=False
+        self,
+        expr: UnaryExpression,
+        df_aliased_col_name_to_real_col_name: Dict[str, str],
+        parse_local_name=False,
     ) -> str:
         if isinstance(expr, Alias):
             quoted_name = quote_name(expr.name)
@@ -410,51 +585,85 @@ class Analyzer:
                 for k, v in self.alias_maps_to_use.items():
                     if v == expr.child.name:
                         self.generated_alias_maps[k] = quoted_name
+
+                for df_alias_dict in df_aliased_col_name_to_real_col_name.values():
+                    for k, v in df_alias_dict.items():
+                        if v == expr.child.name:
+                            df_alias_dict[k] = quoted_name
             return alias_expression(
-                self.analyze(expr.child, parse_local_name), quoted_name
+                self.analyze(
+                    expr.child, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                quoted_name,
             )
         if isinstance(expr, UnresolvedAlias):
-            expr_str = self.analyze(expr.child, parse_local_name)
+            expr_str = self.analyze(
+                expr.child, df_aliased_col_name_to_real_col_name, parse_local_name
+            )
             if parse_local_name:
                 expr_str = expr_str.upper()
             return expr_str
         elif isinstance(expr, Cast):
             return cast_expression(
-                self.analyze(expr.child, parse_local_name), expr.to, expr.try_
+                self.analyze(
+                    expr.child, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                expr.to,
+                expr.try_,
             )
         else:
             return unary_expression(
-                self.analyze(expr.child, parse_local_name),
+                self.analyze(
+                    expr.child, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
                 expr.sql_operator,
                 expr.operator_first,
             )
 
     def binary_operator_extractor(
-        self, expr: BinaryExpression, parse_local_name=False
+        self,
+        expr: BinaryExpression,
+        df_aliased_col_name_to_real_col_name,
+        parse_local_name=False,
     ) -> str:
         if isinstance(expr, BinaryArithmeticExpression):
             return binary_arithmetic_expression(
                 expr.sql_operator,
-                self.analyze(expr.left, parse_local_name),
-                self.analyze(expr.right, parse_local_name),
+                self.analyze(
+                    expr.left, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
+                self.analyze(
+                    expr.right, df_aliased_col_name_to_real_col_name, parse_local_name
+                ),
             )
         else:
             return function_expression(
                 expr.sql_operator,
                 [
-                    self.analyze(expr.left, parse_local_name),
-                    self.analyze(expr.right, parse_local_name),
+                    self.analyze(
+                        expr.left,
+                        df_aliased_col_name_to_real_col_name,
+                        parse_local_name,
+                    ),
+                    self.analyze(
+                        expr.right,
+                        df_aliased_col_name_to_real_col_name,
+                        parse_local_name,
+                    ),
                 ],
                 False,
             )
 
-    def grouping_extractor(self, expr: GroupingSet) -> str:
+    def grouping_extractor(
+        self, expr: GroupingSet, df_aliased_col_name_to_real_col_name
+    ) -> str:
         return self.analyze(
             FunctionExpression(
                 expr.pretty_name.upper(),
                 [c.child if isinstance(c, Alias) else c for c in expr.children],
                 False,
-            )
+            ),
+            df_aliased_col_name_to_real_col_name,
         )
 
     def window_frame_boundary(self, offset: str) -> str:
@@ -464,17 +673,25 @@ class Analyzer:
         except Exception:
             return offset
 
-    def to_sql_avoid_offset(self, expr: Expression, parse_local_name=False) -> str:
+    def to_sql_avoid_offset(
+        self,
+        expr: Expression,
+        df_aliased_col_name_to_real_col_name: Dict[str, str],
+        parse_local_name: bool = False,
+    ) -> str:
         # if expression is a numeric literal, return the number without casting,
         # otherwise process as normal
         if isinstance(expr, Literal) and isinstance(expr.datatype, _NumericType):
             return to_sql_without_cast(expr.value, expr.datatype)
         else:
-            return self.analyze(expr, parse_local_name)
+            return self.analyze(
+                expr, df_aliased_col_name_to_real_col_name, parse_local_name
+            )
 
     def resolve(self, logical_plan: LogicalPlan) -> SnowflakePlan:
         self.subquery_plans = []
         self.generated_alias_maps = {}
+
         result = self.do_resolve(logical_plan)
 
         result.add_aliases(self.generated_alias_maps)
@@ -486,8 +703,16 @@ class Analyzer:
 
     def do_resolve(self, logical_plan: LogicalPlan) -> SnowflakePlan:
         resolved_children = {}
-        for c in logical_plan.children:
-            resolved_children[c] = self.resolve(c)
+        df_aliased_col_name_to_real_col_name: DefaultDict[
+            str, Dict[str, str]
+        ] = defaultdict(dict)
+
+        for c in logical_plan.children:  # post-order traversal of the tree
+            resolved = self.resolve(c)
+            df_aliased_col_name_to_real_col_name.update(
+                resolved.df_aliased_col_name_to_real_col_name
+            )
+            resolved_children[c] = resolved
 
         if isinstance(logical_plan, Selectable):
             # Selectable doesn't have children. It already has the expr_to_alias dict.
@@ -509,53 +734,81 @@ class Analyzer:
                     )
 
             self.alias_maps_to_use = use_maps
-        return self.do_resolve_with_resolved_children(logical_plan, resolved_children)
+
+        res = self.do_resolve_with_resolved_children(
+            logical_plan, resolved_children, df_aliased_col_name_to_real_col_name
+        )
+        res.df_aliased_col_name_to_real_col_name.update(
+            df_aliased_col_name_to_real_col_name
+        )
+        return res
 
     def do_resolve_with_resolved_children(
         self,
         logical_plan: LogicalPlan,
         resolved_children: Dict[LogicalPlan, SnowflakePlan],
+        df_aliased_col_name_to_real_col_name: Dict[str, str],
     ) -> SnowflakePlan:
         if isinstance(logical_plan, SnowflakePlan):
             return logical_plan
 
         if isinstance(logical_plan, TableFunctionJoin):
             return self.plan_builder.join_table_function(
-                self.analyze(logical_plan.table_function),
+                self.analyze(
+                    logical_plan.table_function, df_aliased_col_name_to_real_col_name
+                ),
                 resolved_children[logical_plan.children[0]],
                 logical_plan,
             )
 
         if isinstance(logical_plan, TableFunctionRelation):
             return self.plan_builder.from_table_function(
-                self.analyze(logical_plan.table_function), logical_plan
+                self.analyze(
+                    logical_plan.table_function, df_aliased_col_name_to_real_col_name
+                ),
+                logical_plan,
             )
 
         if isinstance(logical_plan, Lateral):
             return self.plan_builder.lateral(
-                self.analyze(logical_plan.table_function),
+                self.analyze(
+                    logical_plan.table_function, df_aliased_col_name_to_real_col_name
+                ),
                 resolved_children[logical_plan.children[0]],
                 logical_plan,
             )
 
         if isinstance(logical_plan, Aggregate):
             return self.plan_builder.aggregate(
-                list(map(self.to_sql_avoid_offset, logical_plan.grouping_expressions)),
-                list(map(self.analyze, logical_plan.aggregate_expressions)),
+                [
+                    self.to_sql_avoid_offset(expr, df_aliased_col_name_to_real_col_name)
+                    for expr in logical_plan.grouping_expressions
+                ],
+                [
+                    self.analyze(expr, df_aliased_col_name_to_real_col_name)
+                    for expr in logical_plan.aggregate_expressions
+                ],
                 resolved_children[logical_plan.child],
                 logical_plan,
             )
 
         if isinstance(logical_plan, Project):
             return self.plan_builder.project(
-                list(map(self.analyze, logical_plan.project_list)),
+                list(
+                    map(
+                        lambda x: self.analyze(x, df_aliased_col_name_to_real_col_name),
+                        logical_plan.project_list,
+                    )
+                ),
                 resolved_children[logical_plan.child],
                 logical_plan,
             )
 
         if isinstance(logical_plan, Filter):
             return self.plan_builder.filter(
-                self.analyze(logical_plan.condition),
+                self.analyze(
+                    logical_plan.condition, df_aliased_col_name_to_real_col_name
+                ),
                 resolved_children[logical_plan.child],
                 logical_plan,
             )
@@ -574,14 +827,21 @@ class Analyzer:
                 resolved_children[logical_plan.left],
                 resolved_children[logical_plan.right],
                 logical_plan.join_type,
-                self.analyze(logical_plan.condition) if logical_plan.condition else "",
+                self.analyze(
+                    logical_plan.condition, df_aliased_col_name_to_real_col_name
+                )
+                if logical_plan.condition
+                else "",
                 logical_plan,
                 self.session.conf.get("use_constant_subquery_alias", False),
             )
 
         if isinstance(logical_plan, Sort):
             return self.plan_builder.sort(
-                list(map(self.analyze, logical_plan.order)),
+                [
+                    self.analyze(x, df_aliased_col_name_to_real_col_name)
+                    for x in logical_plan.order
+                ],
                 resolved_children[logical_plan.child],
                 logical_plan,
             )
@@ -642,8 +902,12 @@ class Analyzer:
                 logical_plan.child, SnowflakePlan
             ) and isinstance(logical_plan.child.source_plan, Sort)
             return self.plan_builder.limit(
-                self.to_sql_avoid_offset(logical_plan.limit_expr),
-                self.to_sql_avoid_offset(logical_plan.offset_expr),
+                self.to_sql_avoid_offset(
+                    logical_plan.limit_expr, df_aliased_col_name_to_real_col_name
+                ),
+                self.to_sql_avoid_offset(
+                    logical_plan.offset_expr, df_aliased_col_name_to_real_col_name
+                ),
                 resolved_children[logical_plan.child],
                 on_top_of_order_by,
                 logical_plan,
@@ -651,9 +915,16 @@ class Analyzer:
 
         if isinstance(logical_plan, Pivot):
             return self.plan_builder.pivot(
-                self.analyze(logical_plan.pivot_column),
-                [self.analyze(pv) for pv in logical_plan.pivot_values],
-                self.analyze(logical_plan.aggregates[0]),
+                self.analyze(
+                    logical_plan.pivot_column, df_aliased_col_name_to_real_col_name
+                ),
+                [
+                    self.analyze(pv, df_aliased_col_name_to_real_col_name)
+                    for pv in logical_plan.pivot_values
+                ],
+                self.analyze(
+                    logical_plan.aggregates[0], df_aliased_col_name_to_real_col_name
+                ),
                 resolved_children[logical_plan.child],
                 logical_plan,
             )
@@ -662,7 +933,10 @@ class Analyzer:
             return self.plan_builder.unpivot(
                 logical_plan.value_column,
                 logical_plan.name_column,
-                [self.analyze(c) for c in logical_plan.column_list],
+                [
+                    self.analyze(c, df_aliased_col_name_to_real_col_name)
+                    for c in logical_plan.column_list
+                ],
                 resolved_children[logical_plan.child],
                 logical_plan,
             )
@@ -708,7 +982,10 @@ class Analyzer:
                 copy_options=logical_plan.copy_options,
                 validation_mode=logical_plan.validation_mode,
                 column_names=logical_plan.column_names,
-                transformations=[self.analyze(x) for x in logical_plan.transformations]
+                transformations=[
+                    self.analyze(x, df_aliased_col_name_to_real_col_name)
+                    for x in logical_plan.transformations
+                ]
                 if logical_plan.transformations
                 else None,
                 user_schema=logical_plan.user_schema,
@@ -719,7 +996,9 @@ class Analyzer:
             return self.plan_builder.copy_into_location(
                 query=resolved_children[logical_plan.child],
                 stage_location=logical_plan.stage_location,
-                partition_by=self.analyze(logical_plan.partition_by)
+                partition_by=self.analyze(
+                    logical_plan.partition_by, df_aliased_col_name_to_real_col_name
+                )
                 if logical_plan.partition_by
                 else None,
                 file_format_name=logical_plan.file_format_name,
@@ -733,10 +1012,14 @@ class Analyzer:
             return self.plan_builder.update(
                 logical_plan.table_name,
                 {
-                    self.analyze(k): self.analyze(v)
+                    self.analyze(k, df_aliased_col_name_to_real_col_name): self.analyze(
+                        v, df_aliased_col_name_to_real_col_name
+                    )
                     for k, v in logical_plan.assignments.items()
                 },
-                self.analyze(logical_plan.condition)
+                self.analyze(
+                    logical_plan.condition, df_aliased_col_name_to_real_col_name
+                )
                 if logical_plan.condition
                 else None,
                 resolved_children.get(logical_plan.source_data, None),
@@ -746,7 +1029,9 @@ class Analyzer:
         if isinstance(logical_plan, TableDelete):
             return self.plan_builder.delete(
                 logical_plan.table_name,
-                self.analyze(logical_plan.condition)
+                self.analyze(
+                    logical_plan.condition, df_aliased_col_name_to_real_col_name
+                )
                 if logical_plan.condition
                 else None,
                 resolved_children.get(logical_plan.source_data, None),
@@ -757,8 +1042,13 @@ class Analyzer:
             return self.plan_builder.merge(
                 logical_plan.table_name,
                 resolved_children.get(logical_plan.source),
-                self.analyze(logical_plan.join_expr),
-                [self.analyze(c) for c in logical_plan.clauses],
+                self.analyze(
+                    logical_plan.join_expr, df_aliased_col_name_to_real_col_name
+                ),
+                [
+                    self.analyze(c, df_aliased_col_name_to_real_col_name)
+                    for c in logical_plan.clauses
+                ],
                 logical_plan,
             )